[Design] : plan UI 수정 [FEAT] : 앱 실행 시 위치 권한
