+++ app/src/main/java/com/example/tripsync/ui/fragment/plan/planbookmarklist/PlanBoomarkListDialog.kt
@@ -12,7 +12,6 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.tripsync.databinding.FragmentPlanBoomarkListBinding
 import com.example.tripsync.model.Travel
-import com.example.tripsync.ui.fragment.setup.NaverMapFragment.Companion.LOCATION_PERMISSION_REQUEST_CODE
 import com.example.tripsync.ui.fragment.setup.SharedViewModel
 import com.example.tripsync.viewmodel.BookmarkManageViewModel
 import com.example.tripsync.viewmodel.BookmarkManageViewModelFactory
@@ -49,7 +48,6 @@ class PlanBoomarkListDialog : DialogFragment() {
     ): View? {
         _binding = FragmentPlanBoomarkListBinding.inflate(inflater, container, false)
 
-        locationSource = FusedLocationSource(this, LOCATION_PERMISSION_REQUEST_CODE)
         initView()
         return binding.root
     }
