+++ app/src/main/java/com/example/tripsync/ui/fragment/plan/plansearchlist/PlanSearchListAdapter.kt
@@ -66,7 +66,7 @@ class PlanSearchListAdapter(private val itemClickCallBack: (Travel)-> Boolean ) 
             }
 
             planSearchTitle.text = item.title
-            planSearchAddr.text = item.area
+            planSearchAddr.text = item.addr
         }
 
     }
