+++ app/src/main/java/com/example/tripsync/ui/fragment/setup/NaverMapFragment.kt
@@ -29,10 +29,7 @@ import com.naver.maps.map.util.FusedLocationSource
 
 class NaverMapFragment : Fragment(), OnMapReadyCallback {
 
-    private val permissions = arrayOf(
-        Manifest.permission.ACCESS_FINE_LOCATION,
-        Manifest.permission.ACCESS_COARSE_LOCATION
-    )
+
 
 
     private var _binding: NavermapBinding? = null
@@ -46,13 +43,7 @@ class NaverMapFragment : Fragment(), OnMapReadyCallback {
     private val line = PolylineOverlay()
     private val sharedViewModel: SharedViewModel by activityViewModels()
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
 
-        if(!hasPermission()) {
-            ActivityCompat.requestPermissions(requireActivity(), permissions, LOCATION_PERMISSION_REQUEST_CODE)
-        }
-    }
 
 
     override fun onCreateView(
@@ -63,7 +54,6 @@ class NaverMapFragment : Fragment(), OnMapReadyCallback {
         mapView = binding.map
         mapView.onCreate(savedInstanceState)
         mapView.getMapAsync(this)
-        locationSource = FusedLocationSource(this, LOCATION_PERMISSION_REQUEST_CODE)
 
 
         return binding.root
@@ -107,7 +97,6 @@ class NaverMapFragment : Fragment(), OnMapReadyCallback {
                     marker.map = naverMap
                     naverMap.minZoom = 5.0
                     markers.add(marker)
-                    Log.d("map", "Selected locations: ${location.mapX}, ${location.mapY}")
 
                     val cameraPosition = CameraPosition(LatLng(location.mapY, location.mapX), 9.0)
                     naverMap.cameraPosition = cameraPosition
@@ -125,28 +114,19 @@ class NaverMapFragment : Fragment(), OnMapReadyCallback {
 
         val uiSettings = naverMap.uiSettings
         uiSettings.isLocationButtonEnabled = true
-        naverMap.locationTrackingMode = LocationTrackingMode.Face
+        naverMap.locationTrackingMode = LocationTrackingMode.Follow
+
+
 
     }
 
     companion object {
-        const val LOCATION_PERMISSION_REQUEST_CODE = 1000
-
-
         fun newInstance(): NaverMapFragment {
             return NaverMapFragment()
         }
     }
 
-    private fun hasPermission() : Boolean {
-        for (permission in permissions) {
-            if (ContextCompat.checkSelfPermission(requireActivity(), permission)
-                != PackageManager.PERMISSION_GRANTED) {
-                return false
-            }
-        }
-        return true
-    }
+
 
 
 }
