+++ app/src/main/res/layout/fragment_naver_map.xml
@@ -8,7 +8,14 @@
     <com.naver.maps.map.MapView
         android:id="@+id/mapContainer"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+        android:layout_height="match_parent">
+
+        <ImageButton
+            android:id="@+id/naver_map_btn"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:src="@drawable/ic_map"/>
+    </com.naver.maps.map.MapView>
