+++ app/src/main/res/layout/fragment_plan_search_list.xml
@@ -32,17 +32,49 @@
     <androidx.appcompat.widget.SearchView
         android:id="@+id/plan_search_list_search"
         android:layout_width="match_parent"
-        android:layout_height="59dp"
+        android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
         android:layout_marginStart="10dp"
         android:layout_marginEnd="10dp"
         android:background="@drawable/border_user_edit"
+        app:iconifiedByDefault="false"
         android:textSize="15sp"
-        app:closeIcon="@drawable/ic_planbooklist_close"
+        app:searchIcon="@drawable/ic_search2"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/search_iv_logo"
         app:queryHint="가고 싶은 여행지를 검색하세요" />
+
+    <TextView
+        android:id="@+id/plan_search_all"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="4dp"
+        android:layout_marginEnd="64dp"
+        android:text="전체"
+        android:textColor="@color/gray_medium"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/plan_search_list_search" />
+
+    <View
+        android:layout_width="1dp"
+        android:layout_height="15dp"
+        android:background="@color/gray_medium"
+        app:layout_constraintBottom_toTopOf="@+id/plan_search_list_recycler"
+        app:layout_constraintEnd_toStartOf="@+id/plan_search_location"
+        app:layout_constraintStart_toEndOf="@+id/plan_search_all"
+        app:layout_constraintTop_toBottomOf="@+id/plan_search_list_search" />
+
+    <TextView
+        android:id="@+id/plan_search_location"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="4dp"
+        android:layout_marginEnd="10dp"
+        android:text="거리순"
+        android:textColor="@color/gray_medium"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/plan_search_list_search" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/plan_search_list_recycler"
@@ -54,7 +86,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/plan_search_list_search" />
+        app:layout_constraintTop_toBottomOf="@+id/plan_search_all" />
